helper/pyproject.toml
[tool.poetry]
name = "utils"
version = "0.1.0"
description = ""
authors = ["Mateus Santos <santos.mateus@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pynacl = "^1.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
